audio_transcriptor_fernando/requirements.txt
deepgram-sdk
watchdog
customtkinter
python-dotenv
packaging
pyperclip
requests
tkinterdnd2
